# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "D&D 5E Character Creator API",
        "description": "RESTful API for creating and managing D&D 5th Edition (2024 ruleset) characters. Provides endpoints for character CRUD operations and reference data for classes, species, and backgrounds.",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/backgrounds": {
            "get": {
                "tags": [
                    "Backgrounds"
                ],
                "summary": "List all backgrounds",
                "description": "Returns a paginated list of D&D character backgrounds.",
                "operationId": "c75470a6a138e126fcaadb9f75b02bb1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Backgrounds retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Background"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Backgrounds retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/backgrounds/{id}": {
            "get": {
                "tags": [
                    "Backgrounds"
                ],
                "summary": "Get a background",
                "description": "Returns detailed information about a D&D background.",
                "operationId": "d970d72110eaf2c42fc127e0a4e6b2bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Background UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Background retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Background"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Background retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Background not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Background": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    {
                        "properties": {
                            "id": {
                                "description": "Read-only controller for background reference data.",
                                "type": "string",
                                "format": "uuid"
                            },
                            "name": {
                                "type": "string",
                                "example": "Acolyte"
                            },
                            "slug": {
                                "type": "string",
                                "example": "acolyte"
                            },
                            "description": {
                                "type": "string"
                            },
                            "skillProficiencies": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "toolProficiency": {
                                "type": "string",
                                "nullable": true
                            },
                            "toolProficiencyChoices": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "nullable": true
                            },
                            "startingEquipment": {
                                "type": "object"
                            },
                            "startingGold": {
                                "type": "integer",
                                "example": 15
                            },
                            "originFeat": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/Feat"
                                    }
                                ],
                                "nullable": true
                            },
                            "featureName": {
                                "type": "string",
                                "example": "Shelter of the Faithful"
                            },
                            "featureDescription": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Feat": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "name": {
                                "type": "string",
                                "example": "Alert"
                            },
                            "description": {
                                "type": "string"
                            },
                            "prerequisite": {
                                "type": "string",
                                "nullable": true
                            },
                            "category": {
                                "type": "string",
                                "example": "General"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "ErrorResponse": {
                "required": [
                    "success",
                    "error",
                    "meta"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "required": [
                            "code",
                            "message"
                        ],
                        "properties": {
                            "code": {
                                "type": "string",
                                "example": "VALIDATION_ERROR"
                            },
                            "message": {
                                "type": "string",
                                "example": "The given data was invalid."
                            },
                            "details": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "timestamp": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "correlationId": {
                                "type": "string",
                                "format": "uuid"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "currentPage": {
                        "type": "integer",
                        "example": 1
                    },
                    "perPage": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 42
                    },
                    "lastPage": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}